requirements.txt
flask==2.3.3
flask-session==0.5.0
python-dotenv==1.0.0
openai==0.27.1
langchain==0.0.129
faiss-cpu==1.10.0
tiktoken==0.3.0
azure-identity==1.6.0
azure-core==1.33.0
python-dateutil==2.8.2
pytz==2025.2
numpy==1.26.4
pandas==2.2.3
matplotlib==3.6.3
plotly==5.12.0
scipy==1.10.0
scikit-learn==1.2.0
tenacity==8.5.0
